+++ paper/src/main/java/dev/qixils/crowdcontrol/plugin/paper/commands/BlockCommand.java
@@ -3,20 +3,22 @@ package dev.qixils.crowdcontrol.plugin.paper.commands;
 import dev.qixils.crowdcontrol.plugin.paper.FeatureElementCommand;
 import dev.qixils.crowdcontrol.plugin.paper.ImmediateCommand;
 import dev.qixils.crowdcontrol.plugin.paper.PaperCrowdControlPlugin;
+import dev.qixils.crowdcontrol.plugin.paper.utils.ReflectionUtil;
 import dev.qixils.crowdcontrol.socket.Request;
 import dev.qixils.crowdcontrol.socket.Response;
 import lombok.Getter;
 import net.kyori.adventure.text.Component;
-import net.minecraft.world.flag.FeatureFlagSet;
 import org.bukkit.Location;
 import org.bukkit.Material;
 import org.bukkit.block.Block;
-import org.bukkit.craftbukkit.v1_19_R3.util.CraftMagicNumbers;
 import org.bukkit.entity.Player;
 import org.jetbrains.annotations.NotNull;
 import org.jetbrains.annotations.Nullable;
 
 import java.util.List;
+import java.util.Optional;
+
+import static dev.qixils.crowdcontrol.plugin.paper.utils.ReflectionUtil.CB_PACKAGE;
 
 @Getter
 public class BlockCommand extends ImmediateCommand implements FeatureElementCommand {
@@ -41,8 +43,15 @@ public class BlockCommand extends ImmediateCommand implements FeatureElementComm
 	}
 
 	@Override
-	public @NotNull FeatureFlagSet requiredFeatures() {
-		return CraftMagicNumbers.getBlock(material).requiredFeatures();
+	public @NotNull Optional<Object> requiredFeatures() {
+		//return CraftMagicNumbers.getBlock(material).requiredFeatures();
+		return ReflectionUtil.getClazz(CB_PACKAGE + ".util.CraftMagicNumbers").flatMap(clazz -> ReflectionUtil.invokeMethod(
+				(Object) null,
+				clazz,
+				"getBlock",
+				new Class<?>[]{Material.class},
+				material
+		));
 	}
 
 	@Nullable