+++ CHANGELOG.md
@@ -7,7 +7,7 @@ A list of all changes made to the software in reverse chronological order.
 - Added an account autodetection feature which tries to automatically link a streamer's Minecraft
   account and their Crowd Control app session to remove the need to run `/account link`. Certain
   event setups may wish to utilize the `autodetect` config option to disable this feature.
-- Further updates and fixes for Crowd Control 2.0
+- Further updates and fixes for [Crowd Control 2.0](https://beta.crowdcontrol.live/)
 - Increased the height of the Dig Hole effect to destroy partial blocks (i.e. slabs, stairs, etc.)
 - Teleport to X Location effects are now more likely to place a glass block to stand on
 - Backported the Fabric mod to 1.19.2 (experimental) and updated to 1.20 snapshots